Add files via upload
This is a part of the project in which I need to preprocess 4k finger Photos to detect fingerTip. In this I have used YOLOv5s model to detect the fingerTips from those images having different backgrounds. It is feasible with any background e.g. in my dataset it is outdoor , and indoor images.